requirements.txt
#nebula_web
#2019-01-14
atomic==0.7.3
pycparser==2.19
argparse==1.4.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2018.11.29
cffi==1.11.5
click==7.0
elasticsearch==6.3.1
epydoc==3.0.1
greenlet==0.4.14
haigha==0.9.0
ipaddr==2.2.0
pika==0.12.0
psutil==3.3.0
redis==2.10.6
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.21.1
wsgiref==0.1.2
tornado==4.3
phonenumbers==7.2.3
gevent==1.3.7
Jinja2==2.10
MarkupSafe==0.23
PyYAML==3.13
SQLAlchemy==1.2.15
leveldb==0.193
mmh3==2.5.1
#grafana==1.0.0
py2swagger==1.0.0
threathunter_common==1.1.1
babel_python==1.1.1
#tornado_restful_profile_gen==1.0.2
tornado-profile==1.2.0
#nebula_backend==2.0.0
#nebula_web_frontend==3.0.0
nebula_meta==1.1.1
nebula_utils==1.1.1
nebula_parser==1.1.1
nebula_strategy==1.1.1
simplejson==3.16.0
#kerberos==1.2.4 # optional install
aerospike==3.6.0 # global install
futures==3.0.5
xlwt==1.1.2
pylru==1.1.0
pycrypto==2.6.1
Tornado-MySQL==0.5.1
MySQL-python==1.2.5
raven==6.10.0
pytest==4.0.2
pytest-cov==2.6.0
pylint==1.9.4
pudb==2016.2
complexconfig==1.1.1
tornado_profile_gen==1.1.1


